requirements.txt
gradio~=5.34.0
pillow~=11.2.1
torch~=2.7.1
transformers~=4.52.4

transformers
gradio
torch
Pillow
accelerate
